+++ examples/ssr-hydrate-axum/src/app.rs
@@ -18,6 +18,8 @@ pub fn App() -> impl IntoView {
     leptos_fluent! {{
         translations: TRANSLATIONS,
         languages: "./locales/languages.yaml",
+        locales: "./locales",
+        check_translations: "./src/**/*.rs",
         cookie_name: "lang",
         initial_language_from_cookie: true,
         set_language_to_cookie: true,